refactor(transcripts): #230 Update Workspace index
